move userClasses into it own component
